Семинар 17.12.2018

+++ app/src/main/java/info/goodline/cardbag/Photo.java
@@ -5,13 +5,13 @@ import android.widget.ImageView;
 import java.io.Serializable;
 
 public class Photo implements Serializable {
-    private int iconUrl;
+    private long iconUrl;
 
-    public Photo(int iconSources) {
+    public Photo(long iconSources) {
         this.iconUrl = iconSources;
     }
 
-    public int getIconSources() {
+    public long getIconSources() {
         return iconUrl;
     }
 }
